--- packages/components/src/SubHeaderBar/InputTitleSubHeader/index.js
@@ -1,3 +1,0 @@
-import InputTitleSubHeader from './InputTitleSubHeader.component';
-
-export default InputTitleSubHeader;
