+++ packages/components/src/SubHeaderBar/TitleSubHeader/TitleSubHeader.scss
@@ -1,0 +1,66 @@
+@import '~@talend/bootstrap-theme/src/theme/animation';
+
+@mixin ellipsis {
+	white-space: nowrap;
+	text-overflow: ellipsis;
+	overflow: hidden;
+}
+
+@mixin input-text($weight) {
+	color: $black;
+	font-size: $tc-input-subheader-size-medium;
+	font-weight: $weight;
+}
+
+$tc-input-subheader-size-large: 2.4rem !default;
+$tc-input-subheader-size-medium: 1.4rem !default;
+$tc-input-subheader-title-weight: 900 !default;
+$tc-input-subheader-sub-title-weight: 300 !default;
+
+:global(.tc-subheader-details-blink) {
+	@include heartbeat(object-blink);
+}
+
+.tc-subheader-details {
+	display: flex;
+	margin: 0;
+	height: 100%;
+
+	&-icon {
+		align-self: center;
+		width: $tc-input-subheader-size-large;
+		height: $tc-input-subheader-size-large;
+		color: $brand-primary;
+		margin-right: $padding-large;
+	}
+
+	& &-text {
+		display: flex;
+		flex-direction: column;
+		justify-content: center;
+		overflow: hidden;
+		flex: 1 auto;
+		margin-right: 110px;
+		max-width: 90rem;
+
+		&-title {
+			display: inline-flex;
+			align-items: center;
+
+			&-wording,
+			&-wording-button {
+				@include input-text($tc-input-subheader-title-weight);
+				@include ellipsis;
+				text-transform: uppercase;
+				margin: 0;
+			}
+		}
+
+		&-subtitle {
+			@include ellipsis;
+			color: $dark-silver;
+			font-size: $tc-input-subheader-size-medium;
+			font-weight: $tc-input-subheader-sub-title-weight;
+		}
+	}
+}
