+++ packages/components/src/SubHeaderBar/TitleSubHeader/index.js
@@ -1,0 +1,3 @@
+import TitleSubHeader from './TitleSubHeader.component';
+
+export default TitleSubHeader;
